add checking if file already exist in database

+++ controllers/file/uploadFile.js
@@ -6,32 +6,52 @@ import cryptoRandomString from 'crypto-random-string';
 
 const uploadFile = async (req, res) => {
     try {
-        const publicFolderPath = 'D:\\Szkolenia\\Altcom nodeJs\\nodejs-workshop\\miniProjects\\cloud\\Cloud\\public';
-        const extension = fileExtension(req.file.originalname);
-        const randomString = cryptoRandomString({length: 10});
-        const destinationPath = path.join(publicFolderPath, `${randomString}.${extension}`);
-        const temporaryFilePath = req.file.path;
-
-        // zapisanie do bazy i przejście walidacji
         const userId = req.userData.userId;
+        const extension = fileExtension(req.file.originalname);
         const fileName = req.body.fileName ? `${req.body.fileName}.${extension}` : req.file.originalname;
-        const fileInstance = await File.create({
-            name: fileName,
-            path: destinationPath,
-            user_id: userId
+        
+        // walidacja danych
+        const fileExists = await File.findOne({
+            where: {
+                user_id: userId,
+                name: fileName
+            }
         });
 
-        // zapisanie pliku na serwer
-        const fileContent = await fs.readFile(temporaryFilePath);
-        await fs.writeFile(destinationPath, fileContent);
-        await fs.unlink(temporaryFilePath);
+        if(fileExists) {
+            // informacja o konflikcie
+            res.status(409).send({
+                success: false,
+                status: res.status,
+                message: `File with name ${fileName} already exists`
+            });
+        }
+        else {
 
-        res.status(200).send({
-            success: true,
-            status: res.status,
-            data: fileInstance,
-            message: `File successfully uploaded.`
-        });
+            const publicFolderPath = 'D:\\Szkolenia\\Altcom nodeJs\\nodejs-workshop\\miniProjects\\cloud\\Cloud\\public';
+            const randomString = cryptoRandomString({length: 10});
+            const destinationPath = path.join(publicFolderPath, `${randomString}.${extension}`);
+            const temporaryFilePath = req.file.path;
+            // zapisanie do bazy danych
+            const fileInstance = await File.create({
+                name: fileName,
+                path: destinationPath,
+                user_id: userId
+            });
+    
+            // zapisanie pliku na serwer
+            const fileContent = await fs.readFile(temporaryFilePath);
+            await fs.writeFile(destinationPath, fileContent);
+            await fs.unlink(temporaryFilePath);
+            
+            res.status(200).send({
+                success: true,
+                status: res.status,
+                data: fileInstance,
+                message: `File successfully uploaded.`
+            });
+        }
+
     } catch (err) {
         res.status(500).send({
             success: false,
